User Story
I the devoloper added a start menu, end screen, and pause menu along with sounds for all of the buttons that you can click. I have also added 2 new textures for my level one being a red carpet and the other being an indicator for when you are close to a door.
Condition of satisfactiuon
The game functions well and fufills all the requirments for the project